requirements.txt
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-js-asset==2.0.0
django-taggit==2.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.0.1
PyJWT==2.3.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
